+++ drizzle/0006_cloudy_the_santerians.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `streams` ADD `hash` text;--> statement-breakpoint
+CREATE UNIQUE INDEX `uq_streams_hash` ON `streams` (`hash`);