+++ drizzle/0007_light_maverick.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `streams` ADD `size` text;--> statement-breakpoint
+ALTER TABLE `streams` ADD `duration` text;