+++ src/server.ts
@@ -11,6 +11,7 @@ import { buildManifest } from "./lib/manifest.js";
 import { API } from "./utils/constant.js";
 import { ENV } from "./utils/env.js";
 import { Logger } from "./utils/logger.js";
+import { formatStreamTitle } from "./utils/format.js";
 
 const logger = new Logger("SERVER");
 const HOST = "0.0.0.0";